+++ R/acled.api.R
@@ -2,7 +2,7 @@
 #' @name acled.api
 #' @description A small package to access the application programming interface (API) of
 #' the Armed Conflict Location & Event Data
-#' Project (ACLED) at \href{https://www.acleddata.com/}{<https://www.acleddata.com/>}. \cr \cr
+#' Project (ACLED) at \href{https://acleddata.com/}{<https://acleddata.com/>}. \cr \cr
 #' When using this package, you acknowledge that you have read ACLED's terms and conditions
 #' of use, and that you agree with their attribution requirements.
 #' @param regions required numeric vector. Supply one or a vector of multiple integers indicating the geographic region
@@ -19,13 +19,11 @@
 #' observation per event). If set to TRUE, dyadic data is returned.
 #' @param other.query optional character vector. Allows users to add their own ACLED API queries to the
 #' GET call. Note that some query terms require a ? in front.
-#' @param print.data optional logical. When set to FALSE (default), a message is returned that confirms the
-#' successful data access. If set to TRUE, the whole retrieved data set is returned.
 #' @return A data frame containing ACLED events.
 #' @import jsonlite
 #' @import httr
 #' @author Christoph Dworschak \cr Website: \href{https://www.chrisdworschak.com/}{<https://chrisdworschak.com/>}
-#' @references Armed Conflict Location & Event Data Project (ACLED); <https://www.acleddata.com/> \cr
+#' @references Armed Conflict Location & Event Data Project (ACLED); <https://acleddata.com/> \cr
 #' Clionadh Raleigh, Andrew Linke, Havard Hegre and Joakim Karlsen. 2010.
 #' "Introducing ACLED-Armed Conflict Location and Event Data." Journal of Peace Research 47 (5): 651-660.
 #' @examples
@@ -48,21 +46,20 @@ acled.api <- function(
   more.variables = NULL,
   all.variables = FALSE,
   dyadic = FALSE,
-  other.query = NULL,
-  print.data = FALSE){
+  other.query = NULL){
 
   terms <- "read?terms=accept&limit=0"
 
   if (is.null(regions) | !is.numeric(regions) == TRUE) {
     stop("You need to supply a number or a vector of numbers to indicate the geographic regions you wish to retrieve
     (see ACLED's codebook for region codes). For example use: \n
-         acled_data(regions = c(1), start.date = '1995-01-15', end.date = '2005-12-15')", call. = FALSE)
+         acled.api(regions = c(1), start.date = '1995-01-15', end.date = '2005-12-15')", call. = FALSE)
   }
   regions1 <- paste0("&region=", paste(regions, collapse = "|") )
 
   if (is.null(start.date) | is.null(end.date) == TRUE) {
     stop("You need to supply both a start date and an end date. For example use: \n
-         acled_data(regions = c(1), start.date = '1995-01-15', end.date = '2005-12-15')", call. = FALSE)
+         acled.api(regions = c(1), start.date = '1995-01-15', end.date = '2005-12-15')", call. = FALSE)
   }else{
     if ( start.date>end.date ) {
       stop("The start date cannot be larger than the end date.", call. = FALSE)
@@ -102,25 +99,29 @@ acled.api <- function(
   json.content <- jsonlite::fromJSON( httr::content(response, "text", encoding = 'UTF-8'),
                             simplifyVector = FALSE)
   if(!json.content$success){
-    stop(paste0("Request wasn't successful: ", json.content$error$message))
+    stop(paste0("GET request wasn't successful: ", json.content$error$message))
   }
   json.content <- json.content$data
-  if( length(json.content)==0L )return(NULL)
+
+  if( length(json.content)==0L ){
+    message("No data found for this region and time period.")
+    return(NULL)
+  }
+
   acled.matrix <- matrix( unlist(json.content),
                           byrow = T,
-                          nrow = length(json.content))
+                          nrow = length(json.content) )
 
   acled.data <- data.frame(acled.matrix, stringsAsFactors = FALSE)
   names(acled.data) <- names(json.content[[1L]])
 
-  if( is.logical(print.data)==TRUE ){
-    if(print.data==TRUE){ print(acled.data) }else{ message(
-             paste0("Your ACLED data request was successful. ",
-                    length(unique(acled.data$region)), " regions were retrieved for the time starting ",
-                    range(acled.data$event_date)[1], " until ", range(acled.data$event_date)[2], ".")) }
-           }else{
-      stop("The argument 'print.data' requires a logical value.", call. = FALSE)
-    }
+  for(i in 1:ncol(acled.data)){
+    if( all(grepl("^[0-9]+$", acled.data[,i], perl = T))==TRUE ){
+      acled.data[,i] <- as.numeric(acled.data[,i])}
+  }
+
+  message(paste0("Your ACLED data request was successful. Events were retrieved for the period starting ",
+                    range(acled.data$event_date)[1], " until ", range(acled.data$event_date)[2], "."))
 
   return(acled.data)